Change type of indices in Union-Find

+++ src/main/kotlin/bluejam/datastructure/UnionFind.kt
@@ -5,26 +5,26 @@ package bluejam.datastructure
  *
  * Each element is represented by an index (0-indexed).
  */
-class UnionFind(n: Long) {
-    val size = n.toInt()
+class UnionFind(n: Int) {
+    val size = n
 
-    private val parent: Array<Long>
+    private val parent: Array<Int>
     private val cardinality: Array<Long>
 
     init {
-        parent = Array(size) { -1L }
+        parent = Array(size) { -1 }
         cardinality = Array(size) { 1L }
     }
 
     /**
      * Find a root element of a set `x` belongs to.
      */
-    fun find(x: Long): Long {
-        return if (parent[x.toInt()] == -1L) {
+    fun find(x: Int): Int {
+        return if (parent[x] == -1) {
             x
         } else {
-            parent[x.toInt()] = find(parent[x.toInt()])
-            parent[x.toInt()]
+            parent[x] = find(parent[x])
+            parent[x]
         }
     }
 
@@ -34,9 +34,9 @@ class UnionFind(n: Long) {
      *
      * @return If `x` and `y` belong to a same set, return false.
      */
-    fun unite(x: Long, y: Long): Boolean {
-        val u = find(x).toInt()
-        val v = find(y).toInt()
+    fun unite(x: Int, y: Int): Boolean {
+        val u = find(x)
+        val v = find(y)
         if (u == v) {
             return false
         }
@@ -44,10 +44,10 @@ class UnionFind(n: Long) {
         val newSize = cardinality[v] + cardinality[u]
 
         if (cardinality[u] < cardinality[v]) {
-            parent[u] = v.toLong()
+            parent[u] = v
             cardinality[v] = newSize
         } else {
-            parent[v] = u.toLong()
+            parent[v] = u
             cardinality[u] = newSize
         }
 
@@ -59,14 +59,14 @@ class UnionFind(n: Long) {
      *
      * @return `true` if `x` and `y` are in a same set.
      */
-    fun same(x: Long, y: Long): Boolean {
+    fun same(x: Int, y: Int): Boolean {
         return find(x) == find(y)
     }
 
     /**
      * Get a size of a set `x` belongs to.
      */
-    fun getCardinalityOf(x: Long): Long {
-        return cardinality[find(x).toInt()]
+    fun getCardinalityOf(x: Int): Long {
+        return cardinality[find(x)]
     }
 }
